requirements.txt
requests>=2.31.0
pytest
pytest-mock
requests-mock
psycopg2-binary>=2.9
python-dotenv>=1.0

